+++ keras_cv/layers/preprocessing/random_saturation.py
@@ -48,7 +48,7 @@ class RandomSaturation(tf.keras.__internal__.layers.BaseImageAugmentationLayer):
 
     def get_random_transformation(self, image=None, label=None, bounding_box=None):
         del image, label, bounding_box
-        return self.factor.sample()
+        return self.factor()
 
     def augment_image(self, image, transformation=None):
         # Convert the factor range from [0, 1] to [0, +inf]. Note that the
